requirements.txt
py-cord~=2.0.0
python-dotenv
Pillow
motor[srv]
